Add the ability to move "special" lessons to the cell under the date using the -k flag
Added the parsing the argument -k in the main.py to send it parse_json.

In the parser.py it has been added __hash__ function for Note class to add it to set. Added is_kvant flag to specify that this Teacher has "special" lessons. Also "notes" in the Day class has been separate into 2 lists to have the ability to operate them individually.